% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_hoa_doc.R
\docType{data}
\name{cap_hoa}
\alias{cap_hoa}
\title{Homeowner Associations as a vehicle for promoting native urban biodiversity}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39 rows and 13 columns.
}
\source{
{Lerman, S. 2016. Homeowner Associations as a vehicle for promoting native urban biodiversity ver 1. Environmental Data Initiative. https://doi.org/10.6073/pasta/e7fb691d6b4bf6974ece085f22ab63fc (Accessed 2021-05-03).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-cap.640.1}
}
\usage{
cap_hoa
}
\description{
characteristics of HOAs in the greater Phoenix metropolitan area, and corresponding animal (bird or arthropod) and plant diversity within those communities
}
\details{
\describe{
  \item{hoa}{Is the neighborhood part of a HOA? 1=yes, 2=no}
  \item{nat_bird}{Native bird diversity index (Shannon) from PASS bird survey}
  \item{plnt}{Plant diversity index (Shannon) from 2005 Survey 200}
  \item{arthropod}{Arthropod diversity (Shannon) from 2005 Survey 200}
  \item{num_te}{Total number of thin-leaf evergreen trees along 100m transect at Survey 200 plot center}
  }
}
\keyword{datasets}
