% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lef_anole_doc.R
\docType{data}
\name{lef_anole}
\alias{lef_anole}
\title{Add a documentation title here}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 543 rows and 16 columns.
}
\source{
{Enter dataset citation here}
\url{Add link for data package here}
}
\usage{
lef_anole
}
\description{
Add a subtitle / brief summary here
}
\keyword{datasets}
