% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_marsh_doc.R
\docType{data}
\name{pie_marsh}
\alias{pie_marsh}
\title{Aboveground plant biomass and density in control and fertilized plots in a Spartina alterniflora-dominated marsh, Rowley River, Plum Island Ecosystem LTER, MA.}
\format{
A tibble with 32,209 rows and 6 variables:
\describe{
  \item{treatment}{N: fertlized with nitrogen only, P: fertilized with phosphorus only, NP: fertilized with nitrogen and phosphorus, C: control (not fertilized)}
  \item{biomass}{biomass dry weight}
  \item{density}{plant density}
  \item{mean_elev_change}{treatment mean change in height of marsh surface since beginning of experiment (cm)}
  \item{stderr}{treatment standard error}
  }
}
\source{
{Morris, J. 2020. Aboveground plant biomass and density in control and fertilized plots in a Spartina alterniflora-dominated marsh, Rowley River, Plum Island Ecosystem LTER, MA. ver 2. Environmental Data Initiative.}
\url{https://doi.org/10.6073/pasta/d2d045ed569f51d1bf23637b42cf8ea6}
}
\usage{
pie_marsh
}
\description{
Aboveground plant biomass and density is determined non-destructively during the growing season at a Spartina alterniflora-dominated salt marsh on the Rowley River within the Plum Island Ecosystems (PIE) LTER site.
}
\keyword{datasets}
