% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_abundance_doc.R
\docType{data}
\name{sbc_abundance}
\alias{sbc_abundance}
\title{Andrews Experimental Forest vertebrates in Mack Creek}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 74019 rows and 25 columns.
}
\source{
{Gregory, S.V. and I. Arismendi. 2020. Aquatic Vertebrate Population Study in Mack Creek, Andrews Experimental Forest, 1987 to present ver 14. Environmental Data Initiative. https://doi.org/10.6073/pasta/7c78d662e847cdbe33584add8f809165 (Accessed 2021-02-08).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-and.4027.14}
}
\usage{
sbc_abundance
}
\description{
Cutthroat trout and salamander sizes
}
\keyword{datasets}
