% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_hoa_doc.R
\docType{data}
\name{cap_hoa}
\alias{cap_hoa}
\title{Homeowner Associations as a vehicle for promoting native urban biodiversity}
\format{
A tibble with 39 rows and 13 variables:
\describe{
  \item{site}{a character denoting the Survey 200 site name}
  \item{hoa_name}{a character denoting the Legal name of the HOA development}
  \item{muni_1}{a character denoting the name of the Municipality HOA located}
  \item{year}{a number denoting the Year HOA developed}
  \item{hoa}{a number denoting the answer to Is the neighborhood part of a HOA? 1=yes, 2=no}
  \item{nat_bird}{a number denoting the Native bird diversity index (Shannon) from PASS bird survey}
  \item{plnt}{a number denoting the Plant diversity index (Shannon) from 2005 Survey 200}
  \item{arthropod}{a number denoting the Arthropod diversity (Shannon) from 2005 Survey 200}
  \item{income}{a number denoting the Median household income of US Census Block Group}
  \item{yrbuilt}{a number denoting the Average age of housing stock in US Census Block Group}
  \item{x3hadist}{a number denoting the distance to 3.2 ha contiguous desert tract (m) at Survey 200 plot center}
  \item{x1km_imp}{a number denoting the decimal fraction of 1 km plot classified impervious surface at Survey 200 plot center}
  \item{num_te}{a number denoting the total number of thin-leaf evergreen trees along 100m transect at Survey 200 plot center}
  }
}
\source{
{Lerman, S. 2016. Homeowner Associations as a vehicle for promoting native urban biodiversity ver 1. Environmental Data Initiative.}
\url{https://doi.org/10.6073/pasta/e7fb691d6b4bf6974ece085f22ab63fc}
}
\usage{
cap_hoa
}
\description{
characteristics of HOAs in the greater Phoenix metropolitan area, and corresponding animal (bird or arthropod) and plant diversity within those communities
The Shannon diversity index is commonly used to characterize species diversity in a community. It accounts for both abundance and evenness of the species present.
}
\keyword{datasets}
